Agents/AskAmma/requirements.txt
# Core LangChain packages - using compatible versions
langchain>=0.1.0,<0.2.0
langchain-community>=0.0.16,<0.1.0
langchain-core>=0.1.0,<0.2.0
langchain-openai>=0.0.2,<0.1.9

# Web framework and server
fastapi>=0.104.1
uvicorn>=0.24.0

# Data handling and validation
pydantic>=2.5.1

# Environment and configuration
python-dotenv>=1.0.0

# HTTP requests for direct API calls
requests>=2.31.0

# Testing
pytest>=7.4.3